move get/set blocks to AddamsUtils class (for using from all levels configs)
